refactor: project files
